chore: refactor typings import
