+++ scripts/deploy-topos-msg-protocol.ts
@@ -10,8 +10,7 @@ import {
   deployContractConstant,
   predictContractConstant,
 } from './const-addr-deployer'
-import { ToposCore__factory } from '../typechain-types/ethers-v6/factories/contracts/topos-core/ToposCore__factory'
-import { ToposCore } from '../typechain-types/ethers-v6/contracts/topos-core/ToposCore'
+import { ToposCore__factory, ToposCore } from '../typechain-types/ethers-v6'
 
 const main = async function (...args: string[]) {
   const [providerEndpoint, _sequencerPrivateKey] = args
