+++ scripts/test/send-token.ts
@@ -1,7 +1,10 @@
 import { isHexString, JsonRpcProvider, Wallet } from 'ethers'
-import { ERC20Messaging__factory } from '../../typechain-types/ethers-v6/factories/contracts/examples/ERC20Messaging__factory'
-import { BurnableMintableCappedERC20__factory } from '../../typechain-types/ethers-v6/factories/contracts/topos-core/BurnableMintableCappedERC20__factory'
-import { ERC20Messaging } from '../../typechain-types/ethers-v6/contracts/examples/ERC20Messaging'
+
+import {
+  BurnableMintableCappedERC20__factory,
+  ERC20Messaging,
+  ERC20Messaging__factory,
+} from '../../typechain-types/ethers-v6'
 import * as cc from '../../test/topos-core/shared/constants/certificates'
 import * as tc from '../../test/topos-core/shared/constants/tokens'
 import * as testUtils from '../../test/topos-core/shared/utils/common'
