+++ scripts/upgrade-topos-core.ts
@@ -8,8 +8,10 @@ import {
   Wallet,
 } from 'ethers'
 
-import { ToposCore__factory } from '../typechain-types/ethers-v6/factories/contracts/topos-core/ToposCore__factory'
-import { CodeHash__factory } from '../typechain-types/ethers-v6/factories/contracts/topos-core/CodeHash__factory'
+import {
+  ToposCore__factory,
+  CodeHash__factory,
+} from '../typechain-types/ethers-v6'
 
 import toposCoreJSON from '../artifacts/contracts/topos-core/ToposCore.sol/ToposCore.json'
